requirements.txt
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
attrs==26.1.0
bracex==2.6
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
dockerfile-parse==2.0.1
docstring_parser==0.17.0
e2b==2.20.0
e2b-code-interpreter==2.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-anthropic==1.4.0
langchain-core==1.2.26
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
markdown-it-py==4.0.0
mdurl==0.1.2
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
wcmatch==10.1
xxhash==3.6.0
zstandard==0.25.0
